Added context to commands
